+++ src/confighash/config-hash.ts
@@ -3,8 +3,8 @@ import { deflateSync, inflateSync, constants } from 'zlib';
 import * as dictionaryV1 from './dictionary-v1.json';
 
 /**
- * Converts a scrape configuration to a deflated based64 string.
- * @param input - scrape configuration
+ * Converts a project configuration to a deflated based64 string.
+ * @param input - project configuration
  */
 function encode(input: object):string {
   const deflatedIntArr = deflateSync(JSON.stringify(input), { dictionary: Buffer.from(JSON.stringify(dictionaryV1)), level: constants.Z_BEST_COMPRESSION });
@@ -12,8 +12,8 @@ function encode(input: object):string {
 }
 
 /**
- * Converts a deflated based64 string to a scrape configuration.
- * @param deflatedBase64String - scrape configuration
+ * Converts a deflated based64 string to a project configuration.
+ * @param deflatedBase64String - project configuration
  */
 function decode(deflatedBase64String: string) {
   if (!deflatedBase64String || deflatedBase64String.length === 0) return null;